+++ src/test/testData/inspection/MockFieldInitReplaceTest.after.java
@@ -1,0 +1,10 @@
+import org.mockito.Mock;
+
+public class MockFieldInitReplaceTest {
+
+    @Mock
+    public MockObject mock;
+
+    public static class MockObject {
+    }
+}